Master

See merge request CompSci308_2017Fall/slogo_team10!132

+++ src/backend/error_handling/UndefinedColorException.java
@@ -4,7 +4,7 @@ package backend.error_handling;
  * @author Ben Schwennesen
  */
 public class UndefinedColorException extends SyntaxCausedException {
-    public UndefinedColorException(double colorIndex) {
+    public UndefinedColorException(int colorIndex) {
         super(String.valueOf(colorIndex));
     }
 }
